refactor / move things around
